+++ src/context.rs
@@ -60,6 +60,8 @@ mod instance;
 pub mod time;
 mod trigger_tracker;
 
+#[cfg(feature = "reflect")]
+use core::any::type_name;
 use core::{
     any::TypeId,
     cmp::{Ordering, Reverse},
@@ -68,6 +70,8 @@ use core::{
 
 #[cfg(test)]
 use bevy::ecs::system::SystemState;
+#[cfg(feature = "reflect")]
+use bevy::reflect::utility::GenericTypePathCell;
 use bevy::{
     ecs::{
         component::ComponentId,
@@ -160,6 +164,14 @@ impl InputContextAppExt for App {
 
         let _ = self.try_register_required_components::<C, ContextPriority<C>>();
         let _ = self.try_register_required_components::<C, ContextActivity<C>>();
+
+        #[cfg(feature = "reflect")]
+        {
+            self.register_type::<ActionOf<C>>();
+            self.register_type::<Actions<C>>();
+            self.register_type::<ContextActivity<C>>();
+            self.register_type::<ContextPriority<C>>();
+        }
 
         self.add_observer(register::<C, S>)
             .add_observer(unregister::<C, S>)
@@ -703,7 +715,7 @@ fn apply<S: ScheduleLabel>(
 #[cfg_attr(
     feature = "reflect",
     derive(Reflect),
-    reflect(Clone, Component, Default)
+    reflect(Clone, Component, Default, type_path = false)
 )]
 #[component(immutable)]
 pub struct ContextActivity<C> {
@@ -754,6 +766,36 @@ impl<C> Clone for ContextActivity<C> {
 
 impl<C> Copy for ContextActivity<C> {}
 
+#[cfg(feature = "reflect")]
+impl<C: 'static> TypePath for ContextActivity<C> {
+    fn type_path() -> &'static str {
+        static CELL: GenericTypePathCell = GenericTypePathCell::new();
+        CELL.get_or_insert::<Self, _>(|| {
+            format!(
+                concat!(module_path!(), "::ContextActivity<{}>"),
+                type_name::<C>()
+            )
+        })
+    }
+
+    fn short_type_path() -> &'static str {
+        static CELL: GenericTypePathCell = GenericTypePathCell::new();
+        CELL.get_or_insert::<Self, _>(|| format!("ContextActivity<{}>", type_name::<C>()))
+    }
+
+    fn type_ident() -> Option<&'static str> {
+        Some("ContextActivity")
+    }
+
+    fn crate_name() -> Option<&'static str> {
+        Some(module_path!().split(':').next().unwrap())
+    }
+
+    fn module_path() -> Option<&'static str> {
+        Some(module_path!())
+    }
+}
+
 /// Determines the evaluation order of the input context `C` on the entity.
 ///
 /// Used to control how contexts are layered, as some [`Action<C>`]s may consume inputs.
@@ -792,7 +834,7 @@ impl<C> Copy for ContextActivity<C> {}
 #[cfg_attr(
     feature = "reflect",
     derive(Reflect),
-    reflect(Clone, Component, Default)
+    reflect(Clone, Component, Default, type_path = false)
 )]
 #[component(immutable)]
 pub struct ContextPriority<C> {
@@ -824,6 +866,36 @@ impl<C> Clone for ContextPriority<C> {
 }
 
 impl<C> Copy for ContextPriority<C> {}
+
+#[cfg(feature = "reflect")]
+impl<C: 'static> TypePath for ContextPriority<C> {
+    fn type_path() -> &'static str {
+        static CELL: GenericTypePathCell = GenericTypePathCell::new();
+        CELL.get_or_insert::<Self, _>(|| {
+            format!(
+                concat!(module_path!(), "::ContextPriority<{}>"),
+                type_name::<C>()
+            )
+        })
+    }
+
+    fn short_type_path() -> &'static str {
+        static CELL: GenericTypePathCell = GenericTypePathCell::new();
+        CELL.get_or_insert::<Self, _>(|| format!("ContextPriority<{}>", type_name::<C>()))
+    }
+
+    fn type_ident() -> Option<&'static str> {
+        Some("ContextPriority")
+    }
+
+    fn module_path() -> Option<&'static str> {
+        Some(module_path!())
+    }
+
+    fn crate_name() -> Option<&'static str> {
+        Some(module_path!().split(':').next().unwrap())
+    }
+}
 
 /// Associated gamepad for all input contexts on this entity.
 ///